+++ src/frontend/src/config.ts
@@ -1,1 +1,1 @@
-export const APP_VERSION = 'v2.210.0';
+export const APP_VERSION = 'v2.210.1';